main: Correct getopt_long() usage on ARM
getopt_long() returns an int, which is signed.  The existing code casts
this int to a char, which on ARM is unsigned.

Due to how casting works, the resulting value is invalid, and will always
produce an error on ARM.  This prevents seafile-applet from running at all,
as it silently quits before any command line arguments are parsed.

The solution is to change 'char c' to 'int c', which allows for negative
values, and lets getopt_long() work as intended.

Signed-off-by: Sean Cross <xobs@kosagi.com>